+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Features
 + Adding green line around items that are clickable. (Inside the **Not Clickable Items Feature**)
++ Added **Odger waypoint** - Show the Odger waypoint when trophy fishes are in the inventory and no lava rod in hand.
 
 ## Version 0.15 (2023-01-22)